update and add abstract class JavaFaker to avoid duplicated line
