+++ src/main/java/com/rapiddweller/domain/faker/DataFakerGenerator.java
@@ -1,18 +1,13 @@
 package com.rapiddweller.domain.faker;
 
 import com.rapiddweller.benerator.GeneratorContext;
-import com.rapiddweller.benerator.NonNullGenerator;
-import com.rapiddweller.benerator.dataset.DatasetBasedGenerator;
 import com.rapiddweller.benerator.factory.BeneratorExceptionFactory;
-import com.rapiddweller.benerator.wrapper.CompositeGenerator;
-import com.rapiddweller.benerator.wrapper.ProductWrapper;
 import com.rapiddweller.common.exception.IllegalArgumentError;
 import net.datafaker.Faker;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.*;
-import java.util.stream.Collectors;
 
 /*
  * Generate data, base on DataFaker dependency
@@ -20,14 +15,8 @@ import java.util.stream.Collectors;
  *
  */
 
-public class DataFakerGenerator extends CompositeGenerator
-    implements DatasetBasedGenerator, NonNullGenerator {
+public class DataFakerGenerator extends JavaFaker {
 
-  private static final String REGION_NESTING = "com/rapiddweller/dataset/region";
-  private String datasetName;
-  private Locale locale;
-  private final String topic;
-  private final String property;
   private Faker faker;
 
   private final List<String> ignoreTopicsList = Arrays.asList(
@@ -45,112 +34,28 @@ public class DataFakerGenerator extends CompositeGenerator
     this(Locale.getDefault());
   }
 
-  public DataFakerGenerator(Class generatedType) {
-    this(generatedType, Locale.getDefault(), "name", "fullName");
-  }
-
   public DataFakerGenerator(Locale locale) {
-    this(Object.class, locale, "name", "fullName");
+    this(locale, "name", "fullName");
   }
 
   public DataFakerGenerator(String topic, String property) {
-    this(Object.class, Locale.getDefault(), topic, property);
+    this(Locale.getDefault(), topic, property);
   }
 
-  public DataFakerGenerator(Class generatedType, Locale locale, String topic, String property) {
-    super(generatedType);
-    this.locale = locale;
-    this.topic = topic;
-    this.property = property;
-  }
-
-  // properties ------------------------------------------------------------------------------------------------------
-
-  public Locale getLocale() {
-    return locale;
-  }
-
-  public void setLocale(Locale locale) {
-    this.locale = locale;
-  }
-
-  @Override
-  public String getDataset() {
-    return datasetName;
-  }
-
-  // DatasetBasedGenerator interface implementation ------------------------------------------------------------------
-
-  public void setDataset(String datasetName) {
-    this.datasetName = datasetName;
-  }
-
-  @Override
-  public String getNesting() {
-    return REGION_NESTING;
-  }
-
-  @Override
-  public Object generateForDataset(String dataset) {
-    return null;
-  }
-
-  // Generator interface ---------------------------------------------------------------------------------------------
-  @Override
-  public boolean isThreadSafe() {
-    return true;
-  }
-
-  @Override
-  public boolean isParallelizable() {
-    return false;
+  public DataFakerGenerator(Locale locale, String topic, String property) {
+    super(locale, topic, property);
   }
 
   @Override
   public void init(GeneratorContext context) {
-    faker = new Faker(locale);
-
+    faker = new Faker(getLocale());
     errorProperties = getErrorMethod();
 
     super.init(context);
   }
 
   @Override
-  public ProductWrapper generate(ProductWrapper wrapper) {
-    return wrapper.wrap(generate());
-  }
-
-  @Override
-  public Object generate() {
-    return fakerHandler(topic, property);
-  }
-
-  public List<Method> getMethodsIgnoreCase
-      (Class<?> clazz, String methodName) {
-
-    return Arrays.stream(clazz.getMethods())
-        .filter(m -> m.getName().equalsIgnoreCase(methodName))
-        .collect(Collectors.toList());
-  }
-
-  // get method from faker case insensitive
-  public Method getMethodIgnoreCase(String methodName, String fakerPart, Class<?> clazz) {
-
-    List<Method> methods = getMethodsIgnoreCase(clazz, methodName);
-    if (methods.isEmpty()) {
-      throw new IllegalArgumentError("Can't find " + fakerPart + " " + methodName + " in data faker library");
-    }
-
-    //only use method with no parameter
-    for (Method m : methods){
-      if (m.getParameterTypes().length==0){
-        return m;
-      }
-    }
-    throw new IllegalArgumentError(fakerPart + " " + methodName + " is not supported");
-  }
-
-  private Object fakerHandler(String topic, String property) {
+  protected Object fakerHandler(String topic, String property) {
 
     //prevent to access ignoreTopicsList
     if (ignoreTopicsList.contains(topic)){
