+++ src/test/java/com/rapiddweller/domain/faker/FakerGeneratorTest.java
@@ -15,4 +15,14 @@ public class FakerGeneratorTest extends GeneratorTest {
             assertNotNull(tld);
         }
     }
+
+    @Test
+    public void testDataFaker() {
+        DataFakerGenerator generator = new DataFakerGenerator();
+        generator.init(context);
+        for (int i = 0; i < 10; i++) {
+            Object tld = generator.generate();
+            assertNotNull(tld);
+        }
+    }
 }
